+++ R/chapter1.R
@@ -5,85 +5,78 @@
 #'
 #' @param formula Formula for the model
 #' @param data Tibble or data.frame
-#' @param method Regression method, which can be "lm" (default) or "glm"
+#' @param method Regression method, which can be "ols" or "ppml" (default)
 #' @param pair Inter-national fixed effects column (defaults to "pair_id")
 #' @param etfe Exporter time fixed effects column (defaults to "exp_year")
 #' @param itfe Importer time fixed effects column (defaults to "imp_year")
 #' @return A list
 #' @export
 
-tp_summary_app1 <- function(formula, data, method = "lm", pair = "pair_id",
-                                etfe = "exp_year", itfe = "imp_year") {
-  stopifnot(any(method %in% c("lm", "glm")))
+tp_summary_app_1 <- function(formula, data, method = "ppml", pair = "pair_id",
+                            etfe = "exp_year", itfe = "imp_year") {
+  stopifnot(any(method %in% c("ols", "ppml")))
 
   if (!all(class(data) %in% "data.frame")) {
     data <- as.data.frame(data)
   }
 
-  if (method == "lm") {
-    fit <- stats::lm(stats::as.formula(formula), data = data)
-  }
-  if (method == "glm") {
-    fit <- stats::glm(stats::as.formula(formula),
-      family = stats::quasipoisson(link = "log"),
-      data = data,
-      y = FALSE,
-      model = FALSE
-    )
+  if (method == "ols") {
+    fit <- fixest::feols(stats::as.formula(formula),
+                         data = data,
+                         cluster = data[, pair])
   }
 
-  is_ppml <- any(class(fit) %in% "glm")
-
-  contains_etfe <- any(grepl(paste0("^", etfe), names(fit$coefficients)))
-  contains_itfe <- any(grepl(paste0("^", itfe), names(fit$coefficients)))
-
-  vcov_cluster <- sandwich::vcovCL(
-    fit,
-    cluster = data[, pair]
-  )
-
-  coef_test <- lmtest::coeftest(
-    fit,
-    vcov_cluster[
-      which(!grepl(paste0("^", etfe, "|^", itfe), rownames(vcov_cluster))),
-      which(!grepl(paste0("^", etfe, "|^", itfe), rownames(vcov_cluster)))
-    ]
-  )
-
-  coef_test <- broom::tidy(coef_test)
-
-  if (is_ppml) {
-    # reset test: http://personal.lse.ac.uk/tenreyro/reset.do
-    data$predict2 <- (stats::predict(fit))^2 # Get fitted values of the linear index, not of trade
-    form_reset <- stats::update(fit$formula, ~ predict2 + .)
-    fit_reset <- stats::glm(form_reset,
-      family = stats::quasipoisson(link = "log"),
+  if (method == "ppml") {
+    fit <- fixest::fepois(stats::as.formula(formula),
       data = data,
-      y = FALSE,
-      model = FALSE
-    )
-    vcov_cluster_reset <- sandwich::vcovCL(
-      fit_reset,
-      cluster = data[, pair]
-    )
-    res <- lmtest::coeftest(fit_reset, vcov_cluster_reset)
-    res <- res[2, 4]
+      cluster = data[, pair])
+  }
 
-    # r2: http://personal.lse.ac.uk/tenreyro/r2.do
+  is_ppml <- any(fit$method %in% "fepois")
+
+  # regardless of OLS or PPML, the base for the RESET test
+  # is adapted from http://personal.lse.ac.uk/tenreyro/reset.do
+  # we can't use lmtest with fixest
+
+  # For PPML get fitted values of the linear index, not of trade
+  data$predict2 <- (stats::predict(fit, type = "link"))^2
+
+  form_reset <- as.character(stats::update(fit$fml_all$linear, ~ predict2 + .))
+
+  if (length(as.character(fit$fml_all$fixef)) > 0) {
+    form_reset <- paste0(form_reset[2], " ~ ", form_reset[3], " | ",  as.character(fit$fml_all$fixef)[2])
+  } else {
+    form_reset <- paste0(form_reset[2], " ~ ", form_reset[3])
+  }
+
+  if (!is_ppml) {
+    fit_reset <- fixest::feols(stats::as.formula(form_reset),
+                               data = data,
+                               cluster = data[, pair])
+  } else {
+    fit_reset <- fixest::fepois(stats::as.formula(form_reset),
+                                data = data,
+                                cluster = data[, pair])
+  }
+
+  res <- fit_reset$coeftable["predict2", "Pr(>|t|)"]
+
+  r2 <- if (is_ppml) {
+    # Also adapted from http://personal.lse.ac.uk/tenreyro/r2.do
     actual <- as.numeric(data$trade)
     predicted <- as.numeric(fit$fitted.values)
-    r2 <- (stats::cor(actual, predicted, method = "kendall"))^2 # kendall mimics stata
+    (stats::cor(actual, predicted, method = "kendall"))^2 # kendall mimics stata
   } else {
-    res <- lmtest::resettest(fit, power = 2)$p.value
+    fixest::r2(fit, "r2")
   }
 
   return(
     list(
-      tidy_coefficients = coef_test,
+      tidy_coefficients = broom::tidy(fit),
       nobs = nrow(data),
-      rsquared = ifelse(is_ppml, r2, summary(fit)$r.squared),
-      etfe = contains_etfe,
-      itfe = contains_itfe,
+      rsquared = r2,
+      etfe = any(grepl(paste0("^", etfe), fit$fixef_vars)),
+      itfe = any(grepl(paste0("^", itfe), fit$fixef_vars)),
       reset_pval = res
     )
   )
@@ -106,48 +99,29 @@ tp_summary_app1 <- function(formula, data, method = "lm", pair = "pair_id",
 #' @param csfe Country-specific fixed effects (defaults to "intra_pair")
 #' @return A list
 #' @export
-tp_summary_app2 <- function(formula, data, method = "lm",
+tp_summary_app_2 <- function(formula, data, method = "ppml",
                                  pair = "pair_id", etfe = "exp_year",
                                  itfe = "imp_year", dist = "log_dist",
                                  intr = "log_dist_intra", csfe = "intra_pair") {
-  stopifnot(any(method %in% c("lm", "glm")))
+  stopifnot(any(method %in% c("ols", "ppml")))
 
   if (!all(class(data) %in% "data.frame")) {
     data <- as.data.frame(data)
   }
 
-  if (method == "lm") {
-    fit <- stats::lm(stats::as.formula(formula), data = data)
+  if (method == "ols") {
+    fit <- fixest::feols(stats::as.formula(formula), data = data)
   }
-  if (method == "glm") {
-    fit <- stats::glm(stats::as.formula(formula),
-      family = stats::quasipoisson(link = "log"),
+
+  if (method == "ppml") {
+    fit <- fixest::fepois(stats::as.formula(formula),
       data = data,
-      y = FALSE,
-      model = FALSE
+      cluster = data[, pair]
     )
   }
 
-  contains_intr <- any(grepl(paste0("^", intr, "|^", csfe), names(fit$coefficients)))
-  contains_csfe <- any(grepl(paste0("^", csfe), names(fit$coefficients)))
-
-  vcov_cluster <- sandwich::vcovCL(
-    fit,
-    cluster = data[, pair]
-  )
-
-  coef_test <- lmtest::coeftest(
-    fit,
-    vcov_cluster[
-      which(!grepl(paste0("^", etfe, "|^", itfe, "|^", csfe), rownames(vcov_cluster))),
-      which(!grepl(paste0("^", etfe, "|^", itfe, "|^", csfe), rownames(vcov_cluster)))
-    ]
-  )
-
-  coef_test <- broom::tidy(coef_test)
-
   beta_log_dist <- grep(intr,
-    grep(dist, coef_test$term, value = TRUE),
+    grep(dist, names(fit$coefficients), value = TRUE),
     value = TRUE, invert = TRUE
   )
   beta_log_dist <- c(min(beta_log_dist), max(beta_log_dist))
@@ -156,12 +130,16 @@ tp_summary_app2 <- function(formula, data, method = "lm",
   beta1 <- fit$coefficients[min(beta_log_dist)]
   beta2 <- fit$coefficients[max(beta_log_dist)]
 
-  beta_vcov_cluster <- vcov_cluster[
-    which(grepl(paste(beta_log_dist, collapse = "|"), rownames(vcov_cluster))),
-    which(grepl(paste(beta_log_dist, collapse = "|"), rownames(vcov_cluster)))
-  ]
-
   beta_pct_chg <- as.numeric(100 * (beta2 - beta1) / beta1)
+
+  beta_vcov_cluster <- sandwich::vcovCL(
+    fit,
+    cluster = data[, pair]
+  )
+  beta_vcov_cluster <- beta_vcov_cluster[
+    which(grepl(paste(beta_log_dist, collapse = "|"), rownames(beta_vcov_cluster))),
+    which(grepl(paste(beta_log_dist, collapse = "|"), rownames(beta_vcov_cluster)))
+  ]
 
   beta_std_err <- msm::deltamethod(
     ~ 100 * (x2 - x1) / x1,
@@ -173,13 +151,13 @@ tp_summary_app2 <- function(formula, data, method = "lm",
 
   return(
     list(
-      tidy_coefficients = coef_test,
+      tidy_coefficients = broom::tidy(fit),
       nobs = nrow(data),
       pct_chg_log_dist = beta_pct_chg,
       pcld_std_err = beta_std_err,
       pcld_std_err_pval = beta_pval,
-      intr = contains_intr,
-      csfe = contains_csfe
+      intr = any(grepl(paste0("^", intr, "|^", csfe), names(fit$coefficients))),
+      csfe = any(grepl(paste0("^", csfe), names(fit$coefficients)))
     )
   )
 }
@@ -192,7 +170,7 @@ tp_summary_app2 <- function(formula, data, method = "lm",
 #'
 #' @param formula Formula for the model
 #' @param data Tibble or data.frame
-#' @param method Regression method, which can be "lm" (default) or "glm"
+#' @param method Regression method, which can be "ols" or "ppml" (default)
 #' @param pair Inter-national fixed effects column (defaults to "pair_id")
 #' @param pair2 Intra-national fixed effects column (defaults to "pair_id_2")
 #' @param etfe Exporter time fixed effects column (defaults to "exp_year")
@@ -202,73 +180,43 @@ tp_summary_app2 <- function(formula, data, method = "lm",
 #' @param brdr Inter-national borders column (defaults to "intl_brdr")
 #' @return A list
 #' @export
-tp_summary_app3 <- function(formula, data, method = "lm",
+tp_summary_app_3 <- function(formula, data, method = "ppml",
                                  pair = "pair_id", pair2 = "pair_id_2",
                                  etfe = "exp_year", itfe = "imp_year",
                                  dist = "log_dist", intr = "log_dist_intra",
                                  brdr = "intl_brdr") {
-  stopifnot(any(method %in% c("lm", "glm")))
+  stopifnot(any(method %in% c("ols", "ppml")))
 
   if (!all(class(data) %in% "data.frame")) {
     data <- as.data.frame(data)
   }
 
-  if (method == "lm") {
-    fit <- stats::lm(stats::as.formula(formula), data = data)
-  }
-  if (method == "glm") {
-    fit <- stats::glm(stats::as.formula(formula),
-      family = stats::quasipoisson(link = "log"),
+  if (method == "ols") {
+    fit <- fixest::feols(
+      stats::as.formula(formula),
       data = data,
-      y = FALSE,
-      model = FALSE
-    )
+      cluster = data[, pair])
   }
 
-  contains_intr <- any(grepl(
-    paste0("^", intr, "|^", brdr, "|^", pair2),
-    names(fit$coefficients)
-  ))
-
-  vcov_cluster <- sandwich::vcovCL(
-    fit,
-    cluster = data[, pair]
-  )
-
-  vcov_cluster_reduced <- vcov_cluster[
-    which(!grepl(
-      paste0("^", etfe, "|^", itfe, "|^", brdr, "|^", pair2),
-      rownames(vcov_cluster)
-    )),
-    which(!grepl(
-      paste0("^", etfe, "|^", itfe, "|^", brdr, "|^", pair2),
-      rownames(vcov_cluster)
-    ))
-  ]
-
-  if (!is.null(dim(vcov_cluster_reduced))) {
-    coef_test <- lmtest::coeftest(
-      fit,
-      vcov_cluster_reduced
-    )
-
-    coef_test <- broom::tidy(coef_test)
-  } else {
-    coef_test <- broom::tidy(fit) %>%
-      dplyr::filter(
-        !grepl(
-          paste0("^", etfe, "|^", itfe, "|^", brdr, "|^", pair2),
-          term
-        )
-      )
+  if (method == "ppml") {
+    fit <- fixest::fepois(stats::as.formula(formula),
+      data = data,
+      cluster = data[, pair])
   }
+
+  contains_intr <- any(grepl(paste0("^", intr, "|^", brdr, "|^", pair2), names(fit$coefficients)))
 
   beta_rta <- fit$coefficients[grepl("^rta", names(fit$coefficients))]
 
   if (length(beta_rta) > 0) {
-    beta_vcov_cluster <- vcov_cluster[
-      which(grepl(paste(names(beta_rta), collapse = "|"), rownames(vcov_cluster))),
-      which(grepl(paste(names(beta_rta), collapse = "|"), rownames(vcov_cluster)))
+    beta_vcov_cluster <- sandwich::vcovCL(
+      fit,
+      cluster = data[, pair]
+    )
+
+    beta_vcov_cluster <- beta_vcov_cluster[
+      which(grepl(paste(names(beta_rta), collapse = "|"), rownames(beta_vcov_cluster))),
+      which(grepl(paste(names(beta_rta), collapse = "|"), rownames(beta_vcov_cluster)))
     ]
 
     beta_sum <- sum(beta_rta)
@@ -283,7 +231,7 @@ tp_summary_app3 <- function(formula, data, method = "lm",
 
     return(
       list(
-        tidy_coefficients = coef_test,
+        tidy_coefficients = broom::tidy(fit),
         nobs = nrow(data),
         total_rta_effect = beta_sum,
         trta_std_err = beta_std_err,
@@ -294,7 +242,7 @@ tp_summary_app3 <- function(formula, data, method = "lm",
   } else {
     return(
       list(
-        tidy_coefficients = coef_test,
+        tidy_coefficients = broom::tidy(fit),
         nobs = nrow(data),
         intr = contains_intr
       )